--- a/src/META-INF/services/org.daisy.dotify.api.hyphenator.HyphenatorFactoryService
+++ b/src/META-INF/services/org.daisy.dotify.api.hyphenator.HyphenatorFactoryService
@@ -1,2 +1,3 @@
+org.daisy.dotify.hyphenator.impl.SimpleHyphenatorFactoryService
 org.daisy.dotify.hyphenator.impl.LatexHyphenatorFactoryService
 org.daisy.dotify.hyphenator.impl.JHyphenatorFactoryService