--- a/src/org/daisy/dotify/hyphenator/impl/SimpleHyphenator.properties
+++ b/src/org/daisy/dotify/hyphenator/impl/SimpleHyphenator.properties
@@ -1,0 +1,16 @@
+en=hyph_en_GB.dic
+en-AU=hyph_en_US.dic
+en-BZ=hyph_en_US.dic
+en-CA=hyph_en_US.dic
+en-CB=hyph_en_US.dic
+en-GB=hyph_en_GB.dic
+en-IE=hyph_en_US.dic
+en-JM=hyph_en_US.dic
+en-NZ=hyph_en_US.dic
+en-PH=hyph_en_US.dic
+en-TT=hyph_en_US.dic
+en-US=hyph_en_US.dic
+en-ZA=hyph_en_US.dic
+en-ZW=hyph_en_US.dic
+sv=hyph_sv_SE.dic
+sv-SE=hyph_sv_SE.dic