… src/org/daisy/dotify/hyphenator/impl/resource-files/dict/hyph_en_GB.dic
… src/org/daisy/dotify/hyphenator/impl/resource-files/dict/hyph_en_US.dic
… src/org/daisy/dotify/hyphenator/impl/resource-files/dict/hyph_sv_SE.dic
… test/org/daisy/dotify/hyphenator/impl/SimpleHyphenatorSETestOld.java