Hi @bertfrees and @PaulRambags 

I got curious about the hyphenation topic and thought there should be a way to make a simple, readable code for this specific topic. So I spent the weekend creating this SimpleHyphenator.

Moreover, I created this video explaining the concept. (https://youtu.be/XxV6bhCZHn0)

We generally believe this is a better solution that we could improve on going forward, but we don't want it to impact performance. I've run some benchmarks. 

The first-millisecond test is running one book over and over six times. And then, I calculated the mean.

After that, you can see the result of 3 consecutive runs of the regression tests on my personal rig (A bit quicker than usual) 

New Version 1.0.8-SNAPSHOT
```
15803ms
11320ms
12433ms
13291ms
13185ms
11563ms
Mean: 12932ms

BUILD SUCCESSFUL in 49m 12s
BUILD SUCCESSFUL in 49m 15s
BUILD SUCCESSFUL in 49m 20s
```

Old Version 1.0.7
```
11168ms
16631ms
13262ms
14179ms
10486ms
12641ms
Mean: 13061ms

BUILD SUCCESSFUL in 49m 24s
BUILD SUCCESSFUL in 50m 01s
BUILD SUCCESSFUL in 49m 49s
```

My conclusion is that there is no significant difference between the algorithms. Perhaps the old version could be slower or less predictable in the time a book takes to process. But overall there is not any big difference.

Best regards
Daniel