@kalaspuffar 2023-02-28
I forgot to mention that I had converted the Swedish, American English, and Great Brittain English dictionaries. If this solution is approved, I can also convert the other languages so we can migrate to a single solution for all hyphenation, where we can improve the hyphenation code for all languages natively in this repository.