@bertfrees 2023-02-28
Nice :-)

It's nice to have an alternative to https://github.com/daisy/jhyphen. If it behaves the same, performs kind of the same and supports non-standard hyphenation, I could use it instead of jhyphen.

I think it would be good to have a test to prove that it behaves the same as Hyphen. (You mentioned Hunspell in your video, so therefore I assume this is a replacement of jhyphen, not texhyphj.) A way to test this is to hyphenate a big corpus of words using the two implementations that are supposed to be equivalent. By the way, in your existing tests, why do you compare the new implementation with the texhyphj implementation?