@kalaspuffar 2023-03-01
Hi @bertfrees 

I'll answer you in a new comment, as my response is more general than talking about a specific part of the code.

First of I added a couple of test cases for jHyphen, but they are ignored at the moment. For some reason, they don't yield the same result as jTexHyphen or SimpleHyphenator. Not sure why, but there are differences independent of which dictionary I use.

I've tried with the dictionary shipped with my Linux installation and also with the one I generated from the old `tex` format, and both generate differences between what we create in jHyphen and the other two implementations.

I also did some performance benchmarking, but it isn't really fair, as the build failed.
```
11081ms
11804ms
11568ms
12342ms
11412ms
12651ms
Mean:11809ms

BUILD FAILED in 1h 4m 10s
```

As you mentioned in your comment, this functionality could be improved when it comes to performance, and I really hope we will. But my main goal was to make it readable so we could make a joint effort to improve performance later. Currently, it performs as well as the old code so we could only improve :)

With the small change I did in the last commit, we are getting a small performance improvement:
```
12854ms
11486ms
11156ms
13511ms
10315ms
11903ms
Mean:11870ms

BUILD SUCCESSFUL in 49m 50s
```

The reason the full regression test isn't much faster is because of our hyphenation cache, so if we don't have a long-running process working with dotify material, we need to create that each time which entails some upfront work.

The reason I compared this Hyphenator with LatexHyphenator in my original post is that we use that for most languages and then fall back to jHyphen if we have any issues with a specific language. For example, the Russian language crashed on read with LatexHyphenator. 

I'm pretty confident that we can convert the other languages to use the new Hyphenator, and we can handle crash issues much easier when we have all the code in the same place.

Other benefits I can see of not using the older Hyphenators are that we are cross-platform because we have the dictionary files inside of the jar file, and we don't require an external library that might only be available on one platform (jHyphen might not work on Windows or Mac).

Best regards
Daniel