@bertfrees 2023-03-01
I agree there are benefits to using a Java implementation over a C implementation, and I'll definitely consider migrating when we get them to behave the same.

> I've tried with the dictionary shipped with my Linux installation and also with the one I generated from the old tex format, and both generate differences between what we create in jHyphen and the other two implementations.

When you compare the results of different implementations, are the transformations based on the same original TeX patterns (pre-processed with substrings.pl for the jhyphen and SimpleHyphenator implementations)? Also how sure are you that the jtexhyph and SimpleHyphenator yield the same results? Is that based on the small data set that you use in the unit tests?