@kalaspuffar 2023-03-01
Hi @bertfrees 

The jHyphen results are from converting with `substring.pl` and I also tried with my conversion script. And it worked for English. The converted file with `substring.pl` don't handle the exceptions well, but my solution will solve that. 

The failures are more on the Swedish dictionary, which doesn't even handle the first word in the list.

The test ran on jtexhyphen and SimpleHyphenator, both the test cases in the repository and I've also run the full regression tests on both of them without any changes. The baseline of the regression tests was created with jtexhyphen but testing against SimpleHyphenator does not yield any changes.

Some of the later unit test cases were added to test specific problems I found running the regression test.

The best way to test against a larger test set is just to swap the order of the implementations in `org.daisy.dotify.api.hyphenator.HyphenatorFactoryService` and run them against new material. Currently, we only have English and Swedish dictionaries for all Hyphenators. But if you want, I could create Dutch or other ones if you want to test them against your local material.

Or you can use the java class in the repository linked in the video to create different dictionaries. The code is not that extensive, just remove the wrappers and replace all the exception words with specific words with masks of 8's and 9's.

Best regards
Daniel