@bertfrees 2023-03-01
> if you want, I could create Dutch or other ones if you want to test them against your local material

For DAISY Pipeline, I would just like to keep using the exact same dictionaries that I am currently using, only with a different library. DAISY Pipeline includes the dictionaries taken from LibreOffice, except for the Norwegian and German ones which are special ones maintained by NLB and SBS.

Swapping only the library should normally work just fine, because jhyphen and SimpleHyphenator are supposed to support the exact same dictionary file format.

But I guess before testing with a more extended set of test data we need to find out what is going wrong with the Swedish test. Are you really sure you are using the same .dic file? AFAICS `JHyphenatorFactory` uses the one in /usr/share/hyphen and `SimpleHyphenatorFactory` the one in src/org/daisy/dotify/hyphenator/impl/resource-files/dict.